* Refactor interop and add helper types
Move and reorganize native interop code into a dedicated Krypton.Interop area: rename/merge HResult -> NativeFunctions, relocate OS utilities and version types, and extract helpers (BoolExtensions, HResultExtensions, Libraries, SafeModuleHandle, OsVersionInfo). Clean up PlatformInvoke by removing duplicated types and P/Invoke declarations. Set WindowsPlatformVersion in Krypton.Interop.csproj. Also add DesignerSerializationVisibility to VisualEnhancedContextMenu properties, adjust a few TestForm designer sizes/layout calls, and fix file header BOM in LocalWindowsHook. These changes centralize P/Invoke helpers and clarify OS/version plumbing.
