+++ Source/Krypton Components/Krypton.Interop/General/ImageresIconID.cs
